src/main/java/davidch/leetcode/challenges/n83/Solution.java
package davidch.leetcode.challenges.n83;


import davidch.leetcode.challenges.common.linkedlist.ListNode;

public class Solution {
    public ListNode deleteDuplicates(ListNode head) {
        if (head == null) return null;

        var current = head;
        while (current.next != null) {
            if (current.val == current.next.val) {
                current.next = current.next.next;
            } else {
                current = current.next;
            }
        }

        return head;
    }
}
